100%[SUCCESS]

+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: mshereme <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/12 09:44:00 by mshereme          #+#    #+#             */
-/*   Updated: 2023/11/13 17:52:17 by mshereme         ###   ########.fr       */
+/*   Updated: 2023/11/21 09:29:31 by mshereme         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,7 +17,6 @@
 # include <stdlib.h>
 # include <stdarg.h>
 # include <stdint.h>
-# include "printf.h"
 
 # define BASEU "0123456789ABCDEF"
 # define BASEL "0123456789abcdef"
@@ -26,7 +25,7 @@ int	ft_printf(const char *format, ...);
 int	ft_format(char spe, va_list ap);
 int	ft_print_char(int c);
 int	ft_print_str(char *str);
-int	ft_print_ptr(long long unsigned ptr);
+int	ft_print_ptr(uintptr_t ptr);
 int	ft_print_num(long nb);
 int	ft_print_hex(unsigned int nb, char spe);
 int	ft_print_decimal(unsigned int nb);
